Graph plugin bugfix - the render target logic made the implicit assumption that the component had been part of a playing graph at least once before deletion of any associated outbound texture connection. An assumption that's not correct in the case where such a node is pasted then immediately deleted with no playback having occurred in the interim.

+++ plugins/graph.plugin.js
@@ -182,9 +182,15 @@ E2.p.prototype.set_render_target_state = function(on)
 		
 		if(!in_use)
 		{
-			gl.deleteFramebuffer(this.framebuffer);
-			gl.deleteRenderbuffer(this.renderbuffer);
-			this.texture.drop();
+			if(this.framebuffer)
+				gl.deleteFramebuffer(this.framebuffer);
+			
+			if(this.renderbuffer)
+				gl.deleteRenderbuffer(this.renderbuffer);
+				
+			if(this.texture)
+				this.texture.drop();
+			
 			this.framebuffer = null;
 			this.renderbuffer = null;
 			this.texture = null;
